refactor: rename video.ts to videoApi.ts and update imports
